LayoutTests/imported/w3c/web-platform-tests/shadow-dom/reference-target/tentative/dom-mutation-expected.txt



FAIL Changing the ID of the referenced element results in an empty computed label assert_equals: expected "Label 1" but got "Outside the label Label 1 Label 2"
FAIL Removing the referenced element results in an empty computed label assert_equals: expected "Label 1" but got "Outside the label Label 1 Label 2"
FAIL New referenced element prepended to the shadow supercedes the existing label assert_equals: expected "Label 1" but got "Outside the label Label 1 Label 2"
FAIL The existing label supercedes new element (with same id as the existing label) appended to the shadow assert_equals: expected "Label 1" but got "Outside the label Label 1 Label 2"
FAIL Changing the reference target ID updates the computed label assert_equals: expected "Label 1" but got "Outside the label Label 1 Label 2"
FAIL Changing the nested referenceTarget to reference a different element updates the computed label assert_equals: expected "Real Label 1" but got "shadow tree level 1 shadow tree level 2 Real Label 1 Real Label 2"
FAIL Changing the ID of the nested referenced element results in an empty computed label assert_equals: expected "Real Label 1" but got "shadow tree level 1 shadow tree level 2 Real Label 1 Real Label 2"
